+++ src/services/Commerce_CartService.php
@@ -1,7 +1,6 @@
 <?php
 namespace Craft;
 
-use Commerce\Helpers\CommerceDbHelper;
 
 /**
  * Cart service.
@@ -34,7 +33,7 @@ class Commerce_CartService extends BaseApplicationComponent
      */
     public function addToCart($order, $purchasableId, $qty = 1, $note = '', $options = [], &$error = '')
     {
-        CommerceDbHelper::beginStackedTransaction();
+        $transaction = craft()->db->getCurrentTransaction() === null ? craft()->db->beginTransaction() : null;
 
         $isNewLineItem = false;
 
@@ -43,7 +42,10 @@ class Commerce_CartService extends BaseApplicationComponent
         {
             if (!craft()->commerce_orders->saveOrder($order))
             {
-                CommerceDbHelper::rollbackStackedTransaction();
+                if ($transaction !== null)
+                {
+                    $transaction->rollback();
+                }
                 throw new Exception(Craft::t('Error on creating empty cart'));
             }
         }
@@ -87,7 +89,10 @@ class Commerce_CartService extends BaseApplicationComponent
 
                 if (!$event->performAction)
                 {
-                    CommerceDbHelper::rollbackStackedTransaction();
+                    if ($transaction !== null)
+                    {
+                        $transaction->rollback();
+                    }
 
                     return false;
                 }
@@ -103,7 +108,10 @@ class Commerce_CartService extends BaseApplicationComponent
 
                     craft()->commerce_orders->saveOrder($order);
 
-                    CommerceDbHelper::commitStackedTransaction();
+                    if ($transaction !== null)
+                    {
+                        $transaction->commit();
+                    }
 
                     //raising event
                     $event = new Event($this, ['lineItem' => $lineItem, 'order' => $order,]);
@@ -115,11 +123,17 @@ class Commerce_CartService extends BaseApplicationComponent
         }
         catch (\Exception $e)
         {
-            CommerceDbHelper::rollbackStackedTransaction();
+            if ($transaction !== null)
+            {
+                $transaction->rollback();
+            }
             throw $e;
         }
 
-        CommerceDbHelper::rollbackStackedTransaction();
+        if ($transaction !== null)
+        {
+            $transaction->rollback();
+        }
 
         $errors = $lineItem->getAllErrors();
         $error = array_pop($errors);
@@ -463,7 +477,7 @@ class Commerce_CartService extends BaseApplicationComponent
         }
         else
         {
-            CommerceDbHelper::beginStackedTransaction();
+            $transaction = craft()->db->getCurrentTransaction() === null ? craft()->db->beginTransaction() : null;
             try
             {
                 $lineItems = $cart->getLineItems();
@@ -484,13 +498,19 @@ class Commerce_CartService extends BaseApplicationComponent
             }
             catch (\Exception $e)
             {
-                CommerceDbHelper::rollbackStackedTransaction();
+                if ($transaction !== null)
+                {
+                    $transaction->rollback();
+                }
                 CommercePlugin::log($e->getMessage(), LogLevel::Error, true);
 
                 return false;
             }
 
-            CommerceDbHelper::commitStackedTransaction();
+            if ($transaction !== null)
+            {
+                $transaction->commit();
+            }
         }
 
         return true;
@@ -551,7 +571,7 @@ class Commerce_CartService extends BaseApplicationComponent
      */
     public function clearCart(Commerce_OrderModel $cart)
     {
-        CommerceDbHelper::beginStackedTransaction();
+        $transaction = craft()->db->getCurrentTransaction() === null ? craft()->db->beginTransaction() : null;
         try
         {
             craft()->commerce_lineItems->deleteAllLineItemsByOrderId($cart->id);
@@ -559,11 +579,17 @@ class Commerce_CartService extends BaseApplicationComponent
         }
         catch (\Exception $e)
         {
-            CommerceDbHelper::rollbackStackedTransaction();
+            if ($transaction !== null)
+            {
+                $transaction->rollback();
+            }
             throw $e;
         }
 
-        CommerceDbHelper::commitStackedTransaction();
+        if ($transaction !== null)
+        {
+            $transaction->commit();
+        }
     }
 
     /**