+++ src/services/Commerce_OrdersService.php
@@ -6,7 +6,6 @@ use Commerce\Adjusters\Commerce_DiscountAdjuster;
 use Commerce\Adjusters\Commerce_ShippingAdjuster;
 use Commerce\Adjusters\Commerce_TaxAdjuster;
 use Commerce\Helpers\CommerceCurrencyHelper;
-use Commerce\Helpers\CommerceDbHelper;
 
 /**
  * Class Commerce_OrdersService
@@ -245,7 +244,7 @@ class Commerce_OrdersService extends BaseApplicationComponent
         $orderRecord->validate();
         $order->addErrors($orderRecord->getErrors());
 
-        CommerceDbHelper::beginStackedTransaction();
+        $transaction = craft()->db->getCurrentTransaction() === null ? craft()->db->beginTransaction() : null;
 
         try
         {
@@ -257,7 +256,10 @@ class Commerce_OrdersService extends BaseApplicationComponent
                     $orderRecord->id = $order->id;
                     $orderRecord->save(false);
 
-                    CommerceDbHelper::commitStackedTransaction();
+                    if ($transaction !== null)
+                    {
+                        $transaction->commit();
+                    }
 
                     //raising event
                     $event = new Event($this, [
@@ -283,11 +285,17 @@ class Commerce_OrdersService extends BaseApplicationComponent
         }
         catch (\Exception $e)
         {
-            CommerceDbHelper::rollbackStackedTransaction();
+            if ($transaction !== null)
+            {
+                $transaction->rollback();
+            }
             throw $e;
         }
 
-        CommerceDbHelper::rollbackStackedTransaction();
+        if ($transaction !== null)
+        {
+            $transaction->rollback();
+        }
 
         return false;
     }
@@ -560,14 +568,17 @@ class Commerce_OrdersService extends BaseApplicationComponent
         &$error = ''
     )
     {
-        CommerceDbHelper::beginStackedTransaction();
+        $transaction = craft()->db->getCurrentTransaction() === null ? craft()->db->beginTransaction() : null;
         try
         {
             if (!$order->id)
             {
                 if (!$this->saveOrder($order))
                 {
-                    CommerceDbHelper::rollbackStackedTransaction();
+                    if ($transaction !== null)
+                    {
+                        $transaction->rollback();
+                    }
                     throw new Exception(Craft::t('Error on creating empty cart'));
                 }
             }
@@ -613,18 +624,27 @@ class Commerce_OrdersService extends BaseApplicationComponent
                 $order->billingAddressId = $billingAddress->id;
 
                 $this->saveOrder($order);
-                CommerceDbHelper::commitStackedTransaction();
+                if ($transaction !== null)
+                {
+                    $transaction->commit();
+                }
 
                 return true;
             }
         }
         catch (\Exception $e)
         {
-            CommerceDbHelper::rollbackStackedTransaction();
+            if ($transaction !== null)
+            {
+                $transaction->rollback();
+            }
             throw $e;
         }
 
-        CommerceDbHelper::rollbackStackedTransaction();
+        if ($transaction !== null)
+        {
+            $transaction->rollback();
+        }
 
         return false;
     }